test(forward): migrate leak test to dataplane_ut, drop mock/go
The mock/go package was kept alive by a single test verifying that
a stale forward config generation returns memory to the block
allocator. Move the test onto the canonical dataplane_ut harness and
delete both the legacy test directory and the mock/go bindings. The
C-level mock library stays — route_api_test.c still consumes it.
